+++ src/components/DialogForm/DialogForm.jsx
@@ -1,0 +1,48 @@
+import React, { useState } from "react";
+import { makeStyles, useTheme } from "@material-ui/core/styles";
+
+import Button from "@material-ui/core/Button";
+import TextField from "@material-ui/core/TextField";
+import Dialog from "@material-ui/core/Dialog";
+import DialogActions from "@material-ui/core/DialogActions";
+import DialogContent from "@material-ui/core/DialogContent";
+import DialogContentText from "@material-ui/core/DialogContentText";
+import DialogTitle from "@material-ui/core/DialogTitle";
+
+export default function DialogForm(props) {
+  const [open, setOpen] = React.useState(false);
+
+  return (
+    <div>
+      <Dialog
+        open={props.open}
+        onClose={props.handleClose}
+        aria-labelledby="form-dialog-title"
+      >
+        <DialogTitle id="form-dialog-title">Subscribe</DialogTitle>
+        <DialogContent>
+          <DialogContentText>
+            Please fill the following form in order to complete the geofence
+            registration.
+          </DialogContentText>
+          <TextField
+            autoFocus
+            margin="dense"
+            id="name"
+            label="Email Address"
+            type="email"
+            fullWidth
+          />
+        </DialogContent>
+        <DialogActions>
+          <Button onClick={props.handleClose} color="primary">
+            Cancel
+          </Button>
+          <Button onClick={props.handleClose} color="primary">
+            Accept
+          </Button>
+        </DialogActions>
+      </Dialog>
+    </div>
+  );
+}
