+++ src/components/Map/Map.jsx
@@ -9,98 +9,31 @@ import {
 // import { getGeofences,  updateGeofence } from "../../helpers/querys"
 import firebase from "firebase/app";
 import "firebase/firestore";
-import CircularProgress from '@material-ui/core/CircularProgress';
 import Slider from '@material-ui/core/Slider';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
 import Fab from '@material-ui/core/Fab';
 import SaveIcon from '@material-ui/icons/Save';
+import CircularProgress from '@material-ui/core/CircularProgress';
 
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import FormControl from '@material-ui/core/FormControl';
-import InputLabel from '@material-ui/core/InputLabel';
-import Input from '@material-ui/core/Input';
-import Select from '@material-ui/core/Select';
-import Chip from '@material-ui/core/Chip';
-import Button from '@material-ui/core/Button';
-import TextField from '@material-ui/core/TextField';
-
-import MenuItem from '@material-ui/core/MenuItem';
-
-
-const styles = {
-  geofenceForm:{
-    zIndex: 9999,
-    position: "absolute",
-    bottom: 0,
-    left: "50%",
-    right: "50%",
-    marginBottom: 25
-  },
-  paper:{    
-    borderRadius: 50,
-    minWidth: 275,
-    width: 100,
-    height: 65,
-    textAlign:"center",
-    display: "inline",
-    float: "left"
-  },
-  slider:{
-    marginTop: "10px",
-    width: "70%"
-  },
-  addGeofenceButton:{
-    height: 65,
-    width: 65,
-    marginLeft: 15
-  }
-};
+import styles from './styles';
+import DialogForm from '../DialogForm/DialogForm';
 
 const coords = {
   lat: 21.152294,
   lng: -101.711238
 };
 
-const ITEM_HEIGHT = 48;
-const ITEM_PADDING_TOP = 8;
-
-const MenuProps = {
-  PaperProps: {
-    style: {
-      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-      width: 250,
-    },
-  },
-};
-const names = [
-  'Oliver Hansen',
-  'Van Henry',
-  'April Tucker',
-  'Ralph Hubbard',
-  'Omar Alexander',
-  'Carlos Abbott',
-  'Miriam Wagner',
-  'Bradley Wilkerson',
-  'Virginia Andrews',
-  'Kelly Snyder',
-];
-
-
-
 export class MapContainer extends Component {
 
   handleClickOpen = () => {
-    this.setState({open:true});
+    this.setState({dialogForm:{open:true}});
+  };
+  handleClose = () => {
+    this.setState({dialogForm:{open:false}});
   };
 
-  handleClose = () => {
-    this.setState({open:false});
-  };  
+
 
   constructor(props) {
     super(props);
@@ -113,7 +46,9 @@ export class MapContainer extends Component {
       activeMarker: {},
       selectedPlace: {},
       loading: false,
-      open:false,
+      dialogForm:{
+        open:false
+      },
       personName: "",
       error: {
         status: false,
@@ -258,63 +193,16 @@ export class MapContainer extends Component {
               style={styles.addGeofenceButton}
               color="secondary"
               aria-label="add"
-              onClick={()=>{
-                this.setState()
-              }}
+              onClick={this.handleClickOpen}
             >
               <SaveIcon />
             </Fab>
           </div>  
           ):null}
-          <Dialog open={this.state.open} onClose={this.handleClose} aria-labelledby="form-dialog-title">
-          <DialogTitle id="form-dialog-title">Subscribe</DialogTitle>
-            <DialogContent>
-              <DialogContentText>
-                To create the new geofence for a classroom, please complete the following form.
-              </DialogContentText>
-              <TextField
-                autoFocus
-                margin="dense"
-                id="name"
-                label="Courses"
-                type="email"
-                fullWidth
-              />
-              <FormControl >
-                <InputLabel id="demo-mutiple-chip-label">Chip</InputLabel>
-                <Select
-                  labelId="demo-mutiple-chip-label"
-                  id="demo-mutiple-chip"
-                  multiple
-                  value={"hola mundo"}
-                  onChange={this.handleChange}
-                  input={<Input id="select-multiple-chip" />}
-                  renderValue={selected => (
-                    <div >
-                      {selected.map(value => (
-                        <Chip key={value} label={value}  />
-                      ))}
-                    </div>
-                  )}
-                  MenuProps={MenuProps}
-                >
-                  {names.map(name => (
-                    <MenuItem key={name} value={name} >
-                      {name}
-                    </MenuItem>
-                  ))}
-                </Select>
-              </FormControl>
-            </DialogContent>
-            <DialogActions>
-              <Button onClick={console.log("close")} color="primary">
-                Cancel
-              </Button>
-              <Button onClick={console.log("clicked")} color="primary">
-                Accept
-              </Button>
-          </DialogActions>
-        </Dialog>
+          <DialogForm
+            open={this.state.dialogForm.open}
+            handleClose={this.handleClose.bind(this)}
+          />
       </div>
     );
