+++ src/components/Map/styles.jsx
@@ -1,0 +1,31 @@
+
+const styles = {
+    geofenceForm:{
+      zIndex: 9999,
+      position: "absolute",
+      bottom: 0,
+      left: "50%",
+      right: "50%",
+      marginBottom: 25
+    },
+    paper:{    
+      borderRadius: 50,
+      minWidth: 275,
+      width: 100,
+      height: 65,
+      textAlign:"center",
+      display: "inline",
+      float: "left"
+    },
+    slider:{
+      marginTop: "10px",
+      width: "70%"
+    },
+    addGeofenceButton:{
+      height: 65,
+      width: 65,
+      marginLeft: 15
+    }
+  };
+  
+  export default styles;
